+++ spring-cucumber/src/test/java/com/baeldung/ResponseResults.java
@@ -11,23 +11,19 @@ public class ResponseResults {
     private final ClientHttpResponse theResponse;
     private final String body;
 
-    protected ResponseResults(final ClientHttpResponse response) throws IOException {
+    ResponseResults(final ClientHttpResponse response) throws IOException {
         this.theResponse = response;
         final InputStream bodyInputStream = response.getBody();
-        if (null == bodyInputStream) {
-            this.body = "{}";
-        } else {
-            final StringWriter stringWriter = new StringWriter();
-            IOUtils.copy(bodyInputStream, stringWriter);
-            this.body = stringWriter.toString();
-        }
+        final StringWriter stringWriter = new StringWriter();
+        IOUtils.copy(bodyInputStream, stringWriter);
+        this.body = stringWriter.toString();
     }
 
-    protected ClientHttpResponse getTheResponse() {
+    ClientHttpResponse getTheResponse() {
         return theResponse;
     }
 
-    protected String getBody() {
+    String getBody() {
         return body;
     }
 }
