+++ spring-cucumber/src/test/java/com/baeldung/SpringIntegrationTest.java
@@ -1,9 +1,5 @@
 package com.baeldung;
 
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.IntegrationTest;
 import org.springframework.boot.test.SpringApplicationContextLoader;
@@ -12,20 +8,23 @@ import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.web.client.ResponseErrorHandler;
-import org.springframework.web.client.ResponseExtractor;
 import org.springframework.web.client.RestTemplate;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 
 //@RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = SpringDemoApplication.class, loader = SpringApplicationContextLoader.class)
 @WebAppConfiguration
 @IntegrationTest
 public class SpringIntegrationTest {
-    protected static ResponseResults latestResponse = null;
+    static ResponseResults latestResponse = null;
 
     @Autowired
     protected RestTemplate restTemplate;
 
-    protected void executeGet(String url) throws IOException {
+    void executeGet(String url) throws IOException {
         final Map<String, String> headers = new HashMap<>();
         headers.put("Accept", "application/json");
         final HeaderSettingRequestCallback requestCallback = new HeaderSettingRequestCallback(headers);
@@ -39,10 +38,9 @@ public class SpringIntegrationTest {
                 return (new ResponseResults(response));
             }
         });
-
     }
 
-    protected void executePost(String url) throws IOException {
+    void executePost() throws IOException {
         final Map<String, String> headers = new HashMap<>();
         headers.put("Accept", "application/json");
         final HeaderSettingRequestCallback requestCallback = new HeaderSettingRequestCallback(headers);
@@ -53,14 +51,14 @@ public class SpringIntegrationTest {
         }
 
         restTemplate.setErrorHandler(errorHandler);
-        latestResponse = restTemplate.execute(url, HttpMethod.POST, requestCallback, response -> {
-            if (errorHandler.hadError) {
-                return (errorHandler.getResults());
-            } else {
-                return (new ResponseResults(response));
-            }
-        });
-
+        latestResponse = restTemplate
+          .execute("http://localhost:8082/baeldung", HttpMethod.POST, requestCallback, response -> {
+              if (errorHandler.hadError) {
+                  return (errorHandler.getResults());
+              } else {
+                  return (new ResponseResults(response));
+              }
+          });
     }
 
     private class ResponseResultErrorHandler implements ResponseErrorHandler {
