+++ spring-cucumber/src/test/java/com/baeldung/StepDefsIntegrationTest.java
@@ -3,6 +3,7 @@ package com.baeldung;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
+import cucumber.api.java.en.Given;
 import org.springframework.http.HttpStatus;
 
 import cucumber.api.java.en.And;
@@ -10,6 +11,16 @@ import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 
 public class StepDefsIntegrationTest extends SpringIntegrationTest {
+
+    @When("^the client calls /baeldung$")
+    public void the_client_issues_POST_hello() throws Throwable {
+        executePost();
+    }
+
+    @Given("^the client calls /hello$")
+    public void the_client_issues_GET_hello() throws Throwable {
+        executeGet("http://localhost:8082/hello");
+    }
 
     @When("^the client calls /version$")
     public void the_client_issues_GET_version() throws Throwable {
